星状网状差异用例整合

+++ testcases/v2.0/story/java/src/test/java/com/sequoiacm/version/AsyncCacheCurVersionFile1656a.java
@@ -1,15 +1,12 @@
 package com.sequoiacm.version;
 
-import java.io.IOException;
-
-import org.testng.Assert;
+import com.sequoiacm.client.core.*;
+import com.sequoiacm.testcommon.scmutils.ScmFileUtils;
+import org.bson.BSONObject;
 import org.testng.annotations.AfterClass;
 import org.testng.annotations.BeforeClass;
 import org.testng.annotations.Test;
 
-import com.sequoiacm.client.core.ScmFactory;
-import com.sequoiacm.client.core.ScmSession;
-import com.sequoiacm.client.core.ScmWorkspace;
 import com.sequoiacm.client.element.ScmId;
 import com.sequoiacm.client.exception.ScmException;
 import com.sequoiacm.testcommon.ScmInfo;
@@ -20,14 +17,14 @@ import com.sequoiacm.testcommon.WsWrapper;
 import com.sequoiacm.testcommon.scmutils.VersionUtils;
 
 /**
- * test content:update Content of the current scm file, than ayncCache the
- * current version file testlink-case:SCM-1656
- *
+ * @description SCM-1656:异步缓存当前版本文件
  * @author wuyan
- * @Date 2018.06.05
- * @version 1.00
+ * @createDate 2018.06.05
+ * @updateUser ZhangYanan
+ * @updateDate 2021.12.06
+ * @updateRemark
+ * @version v1.0
  */
-
 public class AsyncCacheCurVersionFile1656a extends TestScmBase {
     private static WsWrapper wsp = null;
     private SiteWrapper branSite = null;
@@ -37,13 +34,13 @@ public class AsyncCacheCurVersionFile1656a extends TestScmBase {
     private ScmSession sessionM = null;
     private ScmWorkspace wsM = null;
     private ScmId fileId = null;
-
+    private boolean runSuccess = false;
     private String fileName = "fileVersion1656a";
     private byte[] filedata = new byte[ 1024 * 100 ];
     private byte[] updatedata = new byte[ 1024 * 200 ];
 
     @BeforeClass
-    private void setUp() throws IOException, ScmException {
+    private void setUp() throws ScmException {
         branSite = ScmInfo.getBranchSite();
         rootSite = ScmInfo.getRootSite();
         wsp = ScmInfo.getWs();
@@ -53,6 +50,9 @@ public class AsyncCacheCurVersionFile1656a extends TestScmBase {
         sessionM = TestScmTools.createSession( rootSite );
         wsM = ScmFactory.Workspace.getWorkspace( wsp.getName(), sessionM );
 
+        BSONObject cond = ScmQueryBuilder
+                .start( ScmAttributeName.File.FILE_NAME ).is( fileName ).get();
+        ScmFileUtils.cleanFile( wsp, cond );
         fileId = VersionUtils.createFileByStream( wsM, fileName, filedata );
         VersionUtils.updateContentByStream( wsM, fileId, updatedata );
     }
@@ -72,14 +72,15 @@ public class AsyncCacheCurVersionFile1656a extends TestScmBase {
         // check the historyVersion file only on the rootSite
         SiteWrapper[] expHisSiteList = { rootSite };
         VersionUtils.checkSite( wsA, fileId, historyVersion, expHisSiteList );
+        runSuccess = true;
     }
 
     @AfterClass
-    private void tearDown() {
+    private void tearDown() throws ScmException {
         try {
-            ScmFactory.File.deleteInstance( wsA, fileId, true );
-        } catch ( Exception e ) {
-            Assert.fail( e.getMessage() );
+            if ( runSuccess || TestScmBase.forceClear ) {
+                ScmFactory.File.deleteInstance( wsA, fileId, true );
+            }
         } finally {
             if ( sessionA != null ) {
                 sessionA.close();
@@ -92,13 +93,10 @@ public class AsyncCacheCurVersionFile1656a extends TestScmBase {
 
     private void asyncCacheCurrentVersionFile( int majorVersion )
             throws Exception {
-        // cache
         ScmFactory.File.asyncCache( wsA, fileId, majorVersion, 0 );
-
         int sitenums = 2;
         VersionUtils.waitAsyncTaskFinished( wsM, fileId, majorVersion,
                 sitenums );
-
     }
 
 }
