+++ testcases/v2.0/story/java/src/test/java/com/sequoiacm/version/AsyncCacheCurVersionFile1659.java
@@ -1,17 +1,13 @@
 package com.sequoiacm.version;
 
-import java.io.IOException;
 import java.util.Collection;
-
+import com.sequoiacm.client.core.*;
+import com.sequoiacm.testcommon.scmutils.ScmFileUtils;
+import org.bson.BSONObject;
 import org.testng.Assert;
 import org.testng.annotations.AfterClass;
 import org.testng.annotations.BeforeClass;
 import org.testng.annotations.Test;
-
-import com.sequoiacm.client.core.ScmFactory;
-import com.sequoiacm.client.core.ScmFile;
-import com.sequoiacm.client.core.ScmSession;
-import com.sequoiacm.client.core.ScmWorkspace;
 import com.sequoiacm.client.element.ScmId;
 import com.sequoiacm.client.exception.ScmException;
 import com.sequoiacm.common.ScmFileLocation;
@@ -23,14 +19,14 @@ import com.sequoiacm.testcommon.WsWrapper;
 import com.sequoiacm.testcommon.scmutils.VersionUtils;
 
 /**
- * test content:the historyVersion file in both the rootSite and the branSite,
- * ayncCache the current version file again. testlink-case:SCM-1659
- *
+ * @description SCM-1659:主/分中心已存在历史版本文件，异步缓存当前版本文件
  * @author wuyan
- * @Date 2018.06.05
- * @version 1.00
+ * @createDate 2018.06.05
+ * @updateUser ZhangYanan
+ * @updateDate 2021.12.06
+ * @updateRemark
+ * @version v1.0
  */
-
 public class AsyncCacheCurVersionFile1659 extends TestScmBase {
     private static WsWrapper wsp = null;
     private SiteWrapper branSite = null;
@@ -40,13 +36,13 @@ public class AsyncCacheCurVersionFile1659 extends TestScmBase {
     private ScmSession sessionM = null;
     private ScmWorkspace wsM = null;
     private ScmId fileId = null;
-
+    private boolean runSuccess = false;
     private String fileName = "file1659";
     private byte[] filedata = new byte[ 1024 * 100 ];
     private byte[] updatedata = new byte[ 1024 * 200 ];
 
     @BeforeClass
-    private void setUp() throws IOException, ScmException {
+    private void setUp() throws ScmException {
         branSite = ScmInfo.getBranchSite();
         rootSite = ScmInfo.getRootSite();
         wsp = ScmInfo.getWs();
@@ -56,6 +52,9 @@ public class AsyncCacheCurVersionFile1659 extends TestScmBase {
         sessionM = TestScmTools.createSession( rootSite );
         wsM = ScmFactory.Workspace.getWorkspace( wsp.getName(), sessionM );
 
+        BSONObject cond = ScmQueryBuilder
+                .start( ScmAttributeName.File.FILE_NAME ).is( fileName ).get();
+        ScmFileUtils.cleanFile( wsp, cond );
         fileId = VersionUtils.createFileByStream( wsM, fileName, filedata );
         VersionUtils.updateContentByStream( wsM, fileId, updatedata );
     }
@@ -85,14 +84,15 @@ public class AsyncCacheCurVersionFile1659 extends TestScmBase {
                 secondGetSiteInfo.toString(),
                 "fisrt get siteList:" + firstGetSiteInfo.toString()
                         + " 2nd get siteList:" + secondGetSiteInfo.toString() );
+        runSuccess = true;
     }
 
     @AfterClass
-    private void tearDown() {
+    private void tearDown() throws ScmException {
         try {
-            ScmFactory.File.deleteInstance( wsA, fileId, true );
-        } catch ( Exception e ) {
-            Assert.fail( e.getMessage() );
+            if ( runSuccess || TestScmBase.forceClear ) {
+                ScmFactory.File.deleteInstance( wsA, fileId, true );
+            }
         } finally {
             if ( sessionA != null ) {
                 sessionA.close();
