+++ testcases/v2.0/story/java/src/test/java/com/sequoiacm/version/AsyncCacheHisVersionFile1657.java
@@ -1,15 +1,13 @@
 package com.sequoiacm.version;
 
-import java.io.IOException;
-
+import com.sequoiacm.client.core.*;
+import com.sequoiacm.testcommon.scmutils.ScmFileUtils;
+import org.bson.BSONObject;
 import org.testng.Assert;
 import org.testng.annotations.AfterClass;
 import org.testng.annotations.BeforeClass;
 import org.testng.annotations.Test;
 
-import com.sequoiacm.client.core.ScmFactory;
-import com.sequoiacm.client.core.ScmSession;
-import com.sequoiacm.client.core.ScmWorkspace;
 import com.sequoiacm.client.element.ScmId;
 import com.sequoiacm.client.exception.ScmException;
 import com.sequoiacm.testcommon.ScmInfo;
@@ -20,14 +18,14 @@ import com.sequoiacm.testcommon.WsWrapper;
 import com.sequoiacm.testcommon.scmutils.VersionUtils;
 
 /**
- * test content:update Content of the current scm file, than ayncCache the
- * history version file testlink-case:SCM-1657
- *
+ * @description SCM-1657:异步缓存历史版本文件
  * @author wuyan
- * @Date 2018.06.05
- * @version 1.00
+ * @createDate 2018.06.05
+ * @updateUser ZhangYanan
+ * @updateDate 2021.12.06
+ * @updateRemark
+ * @version v1.0
  */
-
 public class AsyncCacheHisVersionFile1657 extends TestScmBase {
     private static WsWrapper wsp = null;
     private SiteWrapper branSite = null;
@@ -37,13 +35,13 @@ public class AsyncCacheHisVersionFile1657 extends TestScmBase {
     private ScmSession sessionM = null;
     private ScmWorkspace wsM = null;
     private ScmId fileId = null;
-
+    private boolean runSuccess = false;
     private String fileName = "file1657";
     private byte[] filedata = new byte[ 1024 * 100 ];
     private byte[] updatedata = new byte[ 1024 * 200 ];
 
     @BeforeClass
-    private void setUp() throws IOException, ScmException {
+    private void setUp() throws ScmException {
         branSite = ScmInfo.getBranchSite();
         rootSite = ScmInfo.getRootSite();
         wsp = ScmInfo.getWs();
@@ -52,7 +50,9 @@ public class AsyncCacheHisVersionFile1657 extends TestScmBase {
         wsA = ScmFactory.Workspace.getWorkspace( wsp.getName(), sessionA );
         sessionM = TestScmTools.createSession( rootSite );
         wsM = ScmFactory.Workspace.getWorkspace( wsp.getName(), sessionM );
-
+        BSONObject cond = ScmQueryBuilder
+                .start( ScmAttributeName.File.FILE_NAME ).is( fileName ).get();
+        ScmFileUtils.cleanFile( wsp, cond );
         fileId = VersionUtils.createFileByStream( wsM, fileName, filedata );
         VersionUtils.updateContentByStream( wsM, fileId, updatedata );
     }
@@ -74,14 +74,15 @@ public class AsyncCacheHisVersionFile1657 extends TestScmBase {
         // check the currentVersion file only on the rootSite
         SiteWrapper[] expCurSiteList = { rootSite };
         VersionUtils.checkSite( wsA, fileId, currentVersion, expCurSiteList );
+        runSuccess = true;
     }
 
     @AfterClass
-    private void tearDown() {
+    private void tearDown() throws ScmException {
         try {
-            ScmFactory.File.deleteInstance( wsA, fileId, true );
-        } catch ( Exception e ) {
-            Assert.fail( e.getMessage() );
+            if ( runSuccess || TestScmBase.forceClear ) {
+                ScmFactory.File.deleteInstance( wsA, fileId, true );
+            }
         } finally {
             if ( sessionA != null ) {
                 sessionA.close();
@@ -92,13 +93,11 @@ public class AsyncCacheHisVersionFile1657 extends TestScmBase {
         }
     }
 
-    private void asyncCacheCurrentVersionFile( int majorVersion ) {
-        try {
-            // cache
-            ScmFactory.File.asyncCache( wsA, fileId, majorVersion, 0 );
-        } catch ( ScmException e ) {
-            Assert.fail( "asynccAche file fail!" + e.getErrorCode()
-                    + e.getStackTrace() );
-        }
+    private void asyncCacheCurrentVersionFile( int majorVersion )
+            throws Exception {
+        ScmFactory.File.asyncCache( wsA, fileId, majorVersion, 0 );
+        SiteWrapper[] expHisSiteList = { rootSite, branSite };
+        VersionUtils.waitAsyncTaskFinished( wsM, fileId, majorVersion,
+                expHisSiteList.length );
     }
 }
