+++ testcases/v2.0/story/java/src/test/java/com/sequoiacm/version/AsyncTransferCurVersionFile1652a.java
@@ -1,15 +1,11 @@
 package com.sequoiacm.version;
 
-import java.io.IOException;
-
-import org.testng.Assert;
+import com.sequoiacm.client.core.*;
+import com.sequoiacm.testcommon.scmutils.ScmFileUtils;
+import org.bson.BSONObject;
 import org.testng.annotations.AfterClass;
 import org.testng.annotations.BeforeClass;
 import org.testng.annotations.Test;
-
-import com.sequoiacm.client.core.ScmFactory;
-import com.sequoiacm.client.core.ScmSession;
-import com.sequoiacm.client.core.ScmWorkspace;
 import com.sequoiacm.client.element.ScmId;
 import com.sequoiacm.client.exception.ScmException;
 import com.sequoiacm.testcommon.ScmInfo;
@@ -20,41 +16,42 @@ import com.sequoiacm.testcommon.WsWrapper;
 import com.sequoiacm.testcommon.scmutils.VersionUtils;
 
 /**
- * test content:update Content of the current scm file, than ayncTransfer the
- * current version file testlink-case:SCM-1652
- *
+ * @description SCM-1652:异步迁移当前版本文件
  * @author wuyan
- * @Date 2018.06.05
- * @version 1.00
+ * @createDate 2018.06.05
+ * @updateUser ZhangYanan
+ * @updateDate 2021.12.06
+ * @updateRemark
+ * @version v1.0
  */
-
 public class AsyncTransferCurVersionFile1652a extends TestScmBase {
     private static WsWrapper wsp = null;
     private SiteWrapper branSite = null;
     private SiteWrapper rootSite = null;
     private ScmSession sessionA = null;
-    private ScmWorkspace wsA = null;
     private ScmSession sessionM = null;
     private ScmWorkspace wsM = null;
     private ScmId fileId = null;
-
+    private boolean runSuccess = false;
     private String fileName = "fileVersion1652a";
     private byte[] filedata = new byte[ 1024 * 10 ];
     private byte[] updatedata = new byte[ 1024 * 200 ];
 
     @BeforeClass
-    private void setUp() throws IOException, ScmException {
+    private void setUp() throws ScmException {
         branSite = ScmInfo.getBranchSite();
         rootSite = ScmInfo.getRootSite();
         wsp = ScmInfo.getWs();
 
         sessionA = TestScmTools.createSession( branSite );
-        wsA = ScmFactory.Workspace.getWorkspace( wsp.getName(), sessionA );
         sessionM = TestScmTools.createSession( rootSite );
         wsM = ScmFactory.Workspace.getWorkspace( wsp.getName(), sessionM );
 
-        fileId = VersionUtils.createFileByStream( wsA, fileName, filedata );
-        VersionUtils.updateContentByStream( wsA, fileId, updatedata );
+        BSONObject cond = ScmQueryBuilder
+                .start( ScmAttributeName.File.FILE_NAME ).is( fileName ).get();
+        ScmFileUtils.cleanFile( wsp, cond );
+        fileId = VersionUtils.createFileByStream( wsM, fileName, filedata );
+        VersionUtils.updateContentByStream( wsM, fileId, updatedata );
     }
 
     @Test(groups = { "twoSite", "fourSite" })
@@ -65,21 +62,22 @@ public class AsyncTransferCurVersionFile1652a extends TestScmBase {
 
         // check the currentVersion file data and siteinfo
         SiteWrapper[] expCurSiteList = { rootSite, branSite };
-        VersionUtils.checkSite( wsA, fileId, currentVersion, expCurSiteList );
+        VersionUtils.checkSite( wsM, fileId, currentVersion, expCurSiteList );
         VersionUtils.CheckFileContentByStream( wsM, fileName, currentVersion,
                 updatedata );
 
-        // check the historyVersion file only on the branSiteA
-        SiteWrapper[] expHisSiteList = { branSite };
-        VersionUtils.checkSite( wsA, fileId, historyVersion, expHisSiteList );
+        // check the historyVersion file only on the rootSite
+        SiteWrapper[] expHisSiteList = { rootSite };
+        VersionUtils.checkSite( wsM, fileId, historyVersion, expHisSiteList );
+        runSuccess = true;
     }
 
     @AfterClass
-    private void tearDown() {
+    private void tearDown() throws ScmException {
         try {
-            ScmFactory.File.deleteInstance( wsM, fileId, true );
-        } catch ( Exception e ) {
-            Assert.fail( e.getMessage() + e.getStackTrace() );
+            if ( runSuccess || TestScmBase.forceClear ) {
+                ScmFactory.File.deleteInstance( wsM, fileId, true );
+            }
         } finally {
             if ( sessionA != null ) {
                 sessionA.close();
@@ -92,7 +90,8 @@ public class AsyncTransferCurVersionFile1652a extends TestScmBase {
 
     private void asyncTransferCurrentVersionFile( int majorVersion )
             throws Exception {
-        ScmFactory.File.asyncTransfer( wsA, fileId, majorVersion, 0 );
+        ScmFactory.File.asyncTransfer( wsM, fileId, majorVersion, 0,
+                branSite.getSiteName() );
 
         // wait task finished
         int sitenums = 2;
