+++ testcases/v2.0/story/java/src/test/java/com/sequoiacm/version/AsyncTransferCurVersionFile1655.java
@@ -1,17 +1,13 @@
 package com.sequoiacm.version;
 
-import java.io.IOException;
 import java.util.Collection;
-
+import com.sequoiacm.client.core.*;
+import com.sequoiacm.testcommon.scmutils.ScmFileUtils;
+import org.bson.BSONObject;
 import org.testng.Assert;
 import org.testng.annotations.AfterClass;
 import org.testng.annotations.BeforeClass;
 import org.testng.annotations.Test;
-
-import com.sequoiacm.client.core.ScmFactory;
-import com.sequoiacm.client.core.ScmFile;
-import com.sequoiacm.client.core.ScmSession;
-import com.sequoiacm.client.core.ScmWorkspace;
 import com.sequoiacm.client.element.ScmId;
 import com.sequoiacm.client.exception.ScmException;
 import com.sequoiacm.common.ScmFileLocation;
@@ -23,14 +19,14 @@ import com.sequoiacm.testcommon.WsWrapper;
 import com.sequoiacm.testcommon.scmutils.VersionUtils;
 
 /**
- * test content:the historyVersion file in both the rootSite and the branSite,
- * ayncTransfer the current version file. testlink-case:SCM-1655
- *
+ * @description SCM-1655:主/分中心已存在历史版本文件，再次异步迁移当前版本文件
  * @author wuyan
- * @Date 2018.06.05
- * @version 1.00
+ * @createDate 2018.06.05
+ * @updateUser ZhangYanan
+ * @updateDate 2021.12.06
+ * @updateRemark
+ * @version v1.0
  */
-
 public class AsyncTransferCurVersionFile1655 extends TestScmBase {
     private static WsWrapper wsp = null;
     private SiteWrapper branSite = null;
@@ -40,13 +36,13 @@ public class AsyncTransferCurVersionFile1655 extends TestScmBase {
     private ScmSession sessionM = null;
     private ScmWorkspace wsM = null;
     private ScmId fileId = null;
-
+    private boolean runSuccess = false;
     private String fileName = "fileVersion1655";
     private byte[] filedata = new byte[ 1024 * 100 ];
     private byte[] updatedata = new byte[ 1024 * 200 ];
 
     @BeforeClass
-    private void setUp() throws IOException, ScmException {
+    private void setUp() throws ScmException {
         branSite = ScmInfo.getBranchSite();
         rootSite = ScmInfo.getRootSite();
         wsp = ScmInfo.getWs();
@@ -55,7 +51,9 @@ public class AsyncTransferCurVersionFile1655 extends TestScmBase {
         wsA = ScmFactory.Workspace.getWorkspace( wsp.getName(), sessionA );
         sessionM = TestScmTools.createSession( rootSite );
         wsM = ScmFactory.Workspace.getWorkspace( wsp.getName(), sessionM );
-
+        BSONObject cond = ScmQueryBuilder
+                .start( ScmAttributeName.File.FILE_NAME ).is( fileName ).get();
+        ScmFileUtils.cleanFile( wsp, cond );
         fileId = VersionUtils.createFileByStream( wsA, fileName, filedata );
         VersionUtils.updateContentByStream( wsA, fileId, updatedata );
     }
@@ -84,14 +82,15 @@ public class AsyncTransferCurVersionFile1655 extends TestScmBase {
         Assert.assertEquals( getSiteInfo1.toString(), getSiteInfo2.toString(),
                 "fisrt get siteList:" + getSiteInfo1.toString()
                         + " 2nd get siteList:" + getSiteInfo2.toString() );
+        runSuccess = true;
     }
 
     @AfterClass
-    private void tearDown() {
+    private void tearDown() throws ScmException {
         try {
-            ScmFactory.File.deleteInstance( wsM, fileId, true );
-        } catch ( Exception e ) {
-            Assert.fail( e.getMessage() + e.getStackTrace() );
+            if ( runSuccess || TestScmBase.forceClear ) {
+                ScmFactory.File.deleteInstance( wsM, fileId, true );
+            }
         } finally {
             if ( sessionA != null ) {
                 sessionA.close();
@@ -104,7 +103,8 @@ public class AsyncTransferCurVersionFile1655 extends TestScmBase {
 
     private void asyncTransferCurrentVersionFile( int majorVersion )
             throws Exception {
-        ScmFactory.File.asyncTransfer( wsA, fileId, majorVersion, 0 );
+        ScmFactory.File.asyncTransfer( wsA, fileId, majorVersion, 0,
+                rootSite.getSiteName() );
 
         // wait task finished
         int sitenums = 2;
