+++ testcases/v2.0/story/java/src/test/java/com/sequoiacm/version/GetUpdateFileAttr1651.java
@@ -1,16 +1,12 @@
 package com.sequoiacm.version;
 
-import java.io.IOException;
-
+import com.sequoiacm.client.core.*;
+import com.sequoiacm.testcommon.scmutils.ScmFileUtils;
+import org.bson.BSONObject;
 import org.testng.Assert;
 import org.testng.annotations.AfterClass;
 import org.testng.annotations.BeforeClass;
 import org.testng.annotations.Test;
-
-import com.sequoiacm.client.core.ScmFactory;
-import com.sequoiacm.client.core.ScmFile;
-import com.sequoiacm.client.core.ScmSession;
-import com.sequoiacm.client.core.ScmWorkspace;
 import com.sequoiacm.client.element.ScmId;
 import com.sequoiacm.client.exception.ScmException;
 import com.sequoiacm.exception.ScmError;
@@ -22,14 +18,14 @@ import com.sequoiacm.testcommon.WsWrapper;
 import com.sequoiacm.testcommon.scmutils.VersionUtils;
 
 /**
- * test content:write to a file , and updatecontent of the file, gets property
- * information for the specified version file testlink-case:SCM-1651
- *
+ * @description SCM-1651:获取指定版本的文件
  * @author wuyan
- * @Date 2018.06.04
- * @version 1.00
+ * @createDate 2018.06.04
+ * @updateUser ZhangYanan
+ * @updateDate 2021.12.06
+ * @updateRemark
+ * @version v1.0
  */
-
 public class GetUpdateFileAttr1651 extends TestScmBase {
     private static SiteWrapper site = null;
     private static WsWrapper wsp = null;
@@ -41,13 +37,17 @@ public class GetUpdateFileAttr1651 extends TestScmBase {
     private int updateSize = 1024 * 200;
     private byte[] writeData = new byte[ writeSize ];
     private byte[] updateData = new byte[ updateSize ];
+    private boolean runSuccess = false;
 
     @BeforeClass
-    private void setUp() throws IOException, ScmException {
+    private void setUp() throws ScmException {
         site = ScmInfo.getSite();
         wsp = ScmInfo.getWs();
         session = TestScmTools.createSession( site );
         ws = ScmFactory.Workspace.getWorkspace( wsp.getName(), session );
+        BSONObject cond = ScmQueryBuilder
+                .start( ScmAttributeName.File.FILE_NAME ).is( fileName ).get();
+        ScmFileUtils.cleanFile( wsp, cond );
     }
 
     @Test(groups = { "oneSite", "twoSite", "fourSite" })
@@ -65,15 +65,15 @@ public class GetUpdateFileAttr1651 extends TestScmBase {
         getCurrentFileAttr( currentVersion );
         // test c: get file attributes,the version is not exist
         getVersionNoExist();
-
+        runSuccess = true;
     }
 
     @AfterClass
-    private void tearDown() {
+    private void tearDown() throws ScmException {
         try {
-            ScmFactory.File.deleteInstance( ws, fileId, true );
-        } catch ( Exception e ) {
-            Assert.fail( e.getMessage() );
+            if ( runSuccess || TestScmBase.forceClear ) {
+                ScmFactory.File.deleteInstance( ws, fileId, true );
+            }
         } finally {
             if ( session != null ) {
                 session.close();
